feat: check if category is uncategorized before delete

+++ src/usecases/delete-category/delete-category.ts
@@ -14,10 +14,22 @@ export class DeleteCategoryUseCase {
     if (!id) {
       return left(new Error('id not provided'));
     }
+    if (await this.checkIsUncategorized(data)) {
+      return left(new Error('it is not possible to delete this category'));
+    }
     const result = await this.categoryRepository.delete(id);
     if (result.isLeft()) {
       return left(new Error('there was an error in the database operation'));
     }
     return right(id);
+  }
+
+  private async checkIsUncategorized(data: DeleteCategoryData): Promise<boolean> {
+    const categoryOrError = await this.categoryRepository.find({ id: data.id });
+    if (categoryOrError.isLeft()) {
+      return false;
+    }
+    const category = categoryOrError.value[0];
+    return category.description.toLowerCase() === 'não categorizado';
   }
 }
